+++ Maraton/Maraton/templates/formInicio.html
@@ -1,0 +1,18 @@
+{% extends 'base.html' %}
+{% block title %}Saludar{% endblock  %}
+{% block cuerpo %}
+<div style="padding: 1rem 10rem;">
+    <form>
+        <div class="form-group">
+        <label for="calles">Número de calles</label>
+        <input type="number" class="form-control" id="calles">
+        </div>
+        <div class="form-group">
+        <label for="carreras">Número de carreras</label>
+        <input type="number" class="form-control" id="carreras">
+        </div>
+        <button type="submit" class="btn btn-primary">Iniciar</button>
+    </form>
+</div>
+
+{% endblock  %}
