+++ Maraton/Maraton/urls.py
@@ -15,9 +15,10 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from gestionDatos.views import malladoVial
+from gestionDatos.views import malladoVial,inicio
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('inicio/', malladoVial),
+    path('', inicio),
 ]
